Fix naming and demo icons.
