+++ packages/ui/src/services/icon-config/icon-config.ts
@@ -1,4 +1,4 @@
 export { IconMixin } from './IconMixin'
 export { setupIcons } from './setup'
-export { createIconsConfig } from './get-icon-config'
+export { createIconsConfig } from './helpers'
 export { VuesticIconAliases, VuesticIconFonts } from './presets'
